backend/requirements.txt
# Image Recognition Module Dependencies

# Core
openai>=1.0.0
httpx>=0.25.0

# FastAPI Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# OpenAI Agents SDK (for web search)
openai-agents>=0.0.3

# Google Cloud Vision (optional - only needed for Google approach)
google-cloud-vision>=3.0.0

# Image metadata extraction
Pillow>=10.0.0
pillow-heif>=0.13.0  # HEIC/HEIF support for iPhone images
piexif>=1.1.13  # EXIF parsing for HEIC files

# Google Gemini (Google AI Studio) — new SDK for Gemini 3 structured output
google-genai>=1.0.0

# Development
python-dotenv>=1.0.0
